--- a/.changeset/orange-apples-flash.md
+++ b/.changeset/orange-apples-flash.md
@@ -1,0 +1,7 @@
+---
+"@comet/cms-api": patch
+---
+
+Redirects: Improve gql api by preloading PageTree to speed up looking up target pages
+
+also increase the max limit from 100 to 1000