--- a/demo/api/src/db/fixtures/fixtures.console.ts
+++ b/demo/api/src/db/fixtures/fixtures.console.ts
@@ -20,6 +20,7 @@ import slugify from "slugify";
 import { FileUploadsFixtureService } from "./generators/file-uploads-fixture.service";
 import { generateLinks } from "./generators/links.generator";
 import { ManyImagesTestPageFixtureService } from "./generators/many-images-test-page-fixture.service";
+import { RedirectsFixtureService } from "./generators/redirects-fixture.service";
 
 export interface PageTreeNodesFixtures {
     home?: PageTreeNodeInterface;
@@ -49,6 +50,7 @@ export class FixturesConsole {
         @InjectRepository(Link) private readonly linksRepository: EntityRepository<Link>,
         private readonly manyImagesTestPageFixtureService: ManyImagesTestPageFixtureService,
         private readonly fileUploadsFixtureService: FileUploadsFixtureService,
+        private readonly redirectsFixtureService: RedirectsFixtureService,
         private readonly dependenciesService: DependenciesService,
     ) {}
 
@@ -110,6 +112,7 @@ export class FixturesConsole {
 
         node = await this.pageTreeService.createNode(
             {
+                id: "aaa585d3-eca1-47c9-8852-9370817b49ac",
                 name: "Sub",
                 slug: "sub",
                 parentId: node.id,
@@ -277,6 +280,8 @@ export class FixturesConsole {
         }
 
         await this.fileUploadsFixtureService.generateFileUploads();
+
+        await this.redirectsFixtureService.generateRedirects();
 
         await this.dependenciesService.createViews();