--- a/demo/api/src/db/migrations/Migration20241015162102.ts
+++ b/demo/api/src/db/migrations/Migration20241015162102.ts
@@ -1,0 +1,12 @@
+import { Migration } from '@mikro-orm/migrations';
+
+export class Migration20241015162102 extends Migration {
+
+  async up(): Promise<void> {
+    this.addSql('CREATE INDEX "Redirect_scope_domain" on "Redirect" ("scope_domain")');
+  }
+
+  async down(): Promise<void> {
+    this.addSql('DROP INDEX "Redirect_scope_domain"');
+  }
+}