--- a/demo/api/src/redirects/dto/redirect-scope.ts
+++ b/demo/api/src/redirects/dto/redirect-scope.ts
@@ -1,4 +1,4 @@
-import { Embeddable, Property } from "@mikro-orm/core";
+import { Embeddable, Index, Property } from "@mikro-orm/core";
 import { Field, InputType, ObjectType } from "@nestjs/graphql";
 import { IsString } from "class-validator";
 
@@ -7,6 +7,7 @@ import { IsString } from "class-validator";
 @InputType("RedirectScopeInput") // name must not be changed in the app
 // @TODO: disguise @ObjectType("RedirectScope") and @InputType("RedirectScopeInput") decorators under a custom decorator: f.i. @RedirectScope
 export class RedirectScope {
+    @Index() // this does nothing, migration has to be created manually as the entity is in library
     @Property({ columnType: "text" })
     @Field()
     @IsString()