--- a/packages/api/cms-api/src/redirects/redirects.module.ts
+++ b/packages/api/cms-api/src/redirects/redirects.module.ts
@@ -44,11 +44,13 @@ export class RedirectsModule {
             useValue: linkBlock,
         };
 
+        const mikroOrmModule = MikroOrmModule.forFeature([Redirect]);
+
         return {
             module: RedirectsModule,
-            imports: [MikroOrmModule.forFeature([Redirect])],
+            imports: [mikroOrmModule],
             providers: [RedirectsResolver, RedirectsDependenciesResolver, RedirectsService, linkBlockProvider, ImportRedirectsConsole],
-            exports: [RedirectsService],
+            exports: [RedirectsService, REDIRECTS_LINK_BLOCK, mikroOrmModule],
         };
     }
 }