Large number of redirects can cause issues in real projects. Help on improving the situation by making the API a bit faster.(that alone won't be the solution).

- Add fixtures that create 7000 redirects (for this the additional exports are needed)
- Add a relatively simple performance optimization in redirects resolver: Call `this.pageTreeReadApi.preloadNodes(scope);` to load the whole page tree into memory, as most redirects will access it (when they are internal)
    - Local tests showed (when loading 100 redirects) ~250ms (before) vs ~100ms (after) (!)
- (Demo only) Add an index on `Redirect.scope_domain`
   - No difference in Demo performance, but much needed in "real" multidomain applications
   - Unfortunately this won't be generated by MikroORM, as the redirect entity is in library only
- And increase the maximum limit for paginated redirects query from 100 to 1000
    - Not much improvement, but still ~4.8s (before) vs ~3.2s (after) (both with preloadNodes)