@johnnyomair 2024-10-16
> unfortunately this won't be generated by mikro-orm, as Redirect entity is in library only

Should we move the entity to the application, like we did for DAM and the page tree?