@nsams 2024-10-16
> Should we move the entity to the application, like we did for DAM and the page tree?

yes, I thought about that too