@johnnyomair 2024-10-16
> yes, I thought about that too

Created a task for this: https://vivid-planet.atlassian.net/browse/COM-1183.